+++ src/live.rs
@@ -268,9 +268,10 @@ pub struct Subscription {
     /// The symbols of the instruments to subscribe to.
     #[builder(setter(into))]
     pub symbols: Symbols,
-    /// The schema of data to subscribe to.
+    /// The data record schema of data to subscribe to.
     pub schema: Schema,
-    /// The symbology type of symbols in [`symbols`](Self::symbols).
+    /// The symbology type of the symbols in [`symbols`](Self::symbols).
+    #[builder(default = SType::RawSymbol)]
     pub stype_in: SType,
     /// If specified, requests available data since that time. When `None`,
     /// only real-time data is sent.